Moved definition of * (bell; change) into change.pxd
This was originally defined on the bell class to work around import problems,
but this was not ideal as it is defined within change.h.
Possible future module separations would have failed if bell.h was not included when required.

Using "from change cimport *" now imports the operator correctly, meaning we can move it to join other change declarations.